+++ sensor/scanner/openvas/openvas_test.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
-	"sync"
 	"testing"
 	"time"
 )
@@ -40,7 +39,7 @@ type helperLongCommander struct {
 }
 
 func (exe helperLongCommander) Command(name string, arg ...string) *exec.Cmd {
-	cs := []string{"-test.run=TestCommandEndless", "--", name}
+	cs := []string{"-test.run=TestCommandLong", "--", name}
 	cs = append(cs, arg...)
 	cmd := exec.Command(os.Args[0], cs...)
 	cmd.Env = []string{"GO_TEST_PROCESS=1"}
@@ -49,13 +48,11 @@ func (exe helperLongCommander) Command(name string, arg ...string) *exec.Cmd {
 
 // TestCommandEndless is not a real test. It is only used as a helper process to
 // simulate long running programm such as a scan in openvas
-func TestCommandEndless(t *testing.T) {
+func TestCommandLong(t *testing.T) {
 	if os.Getenv("GO_TEST_PROCESS") != "1" {
 		return
 	}
-	for {
-		time.Sleep(time.Second)
-	}
+	time.Sleep(time.Second)
 }
 
 // helperFailCommander creates a Command to execute a programm with a exit code 1
@@ -127,90 +124,81 @@ func TestCommandSettings(t *testing.T) {
 
 // TestStartStopScanSudo tests the procedure of creating an openvas process and stopping it with sudo privileges
 func TestStartStopScanSudo(t *testing.T) {
-	// Create Process List for test
-	var processes = ProcessList{
-		procs: make(map[string]*os.Process),
-		mutex: &sync.Mutex{},
-	}
-
-	// Get sudo rights
+	// OpenVASScanner instance
+	ovas := NewOpenVASScanner()
 	sudo := IsSudo(helperShortCommander{})
+
+	// Test for sudo rights
 	if !sudo {
 		t.Fatalf("Error: Sudo is %t, but should be %t", sudo, true)
 	}
 
 	// Start scan
-	if err := StartScan("foo", 10, sudo, helperLongCommander{}, processes); err != nil {
+	if err := ovas.StartScan("foo", 10, sudo, helperLongCommander{}); err != nil {
 		t.Fatalf("Error: Cannot run StartScan: %s", err)
 	}
 
 	// Check if process is added to the list
-	if _, ok := processes.procs["foo"]; !ok {
+	if _, ok := ovas.procs["foo"]; !ok {
 		t.Fatal("Error: Unable to locate Process")
 	}
 
 	// Stop Scan
-	if err := StopScan("foo", sudo, helperShortCommander{}, processes); err != nil {
+	if err := ovas.StopScan("foo", sudo, helperShortCommander{}); err != nil {
 		t.Fatalf("Error: Unable to stop process: %s", err)
 	}
 
 	// Check if process is removed from the list
-	if _, ok := processes.procs["foo"]; ok {
+	if _, ok := ovas.procs["foo"]; ok {
 		t.Fatalf("Error: Process still in process list")
 	}
 }
 
 // TestNonSudoStopScanFail tests if it fails to stop a scan when there is no scan to stop
 func TestNonSudoStopScanFail(t *testing.T) {
-	// Create Process List for test
-	var processes = ProcessList{
-		procs: make(map[string]*os.Process),
-		mutex: &sync.Mutex{},
-	}
+	// OpenVASScanner instance
+	ovas := NewOpenVASScanner()
+	sudo := IsSudo(helperFailCommander{})
 
 	// Test if sudo is unavailable
-	sudo := IsSudo(helperFailCommander{})
 	if sudo {
 		t.Fatalf("Error: Sudo is %t, but should be %t", sudo, false)
 	}
 
 	// Stop Scan should fail because ther is no process
-	if err := StopScan("foo", sudo, helperShortCommander{}, processes); err == nil {
+	if err := ovas.StopScan("foo", sudo, helperShortCommander{}); err == nil {
 		t.Fatalf("Error: Should be unable to successfully stop scan")
 	}
 }
 
 // TestScanFinishedSuccess tests if it can mark a scan as finished
 func TestScanFinishedSuccess(t *testing.T) {
-	// Create Process List for test
-	var processes = ProcessList{
-		procs: make(map[string]*os.Process),
-		mutex: &sync.Mutex{},
-	}
+	// OpenVASScanner instance
+	ovas := NewOpenVASScanner()
 
-	processes.addProcess("foo", nil)
+	ovas.addProcess("foo", nil)
 
-	if err := ScanFinished("foo", processes); err != nil {
+	if err := ovas.ScanFinished("foo"); err != nil {
 		t.Fatalf("Error: Unable to finish scan: %s", err)
 	}
 }
 
 // TestScanFinishedFail tests if marking a scan as finished fails if there is no scan to finish
 func TestScanFinishedFail(t *testing.T) {
-	// Create Process List for test
-	var processes = ProcessList{
-		procs: make(map[string]*os.Process),
-		mutex: &sync.Mutex{},
-	}
+	// OpenVASScanner instance
+	ovas := NewOpenVASScanner()
 
-	if err := ScanFinished("foo", processes); err == nil {
+	if err := ovas.ScanFinished("foo"); err == nil {
 		t.Fatalf("Error: ScanFinished should return an error")
 	}
 }
 
 // TestGetVersion tests if the information getting from the openvas version is extracted correctly
 func TestGetVersion(t *testing.T) {
-	ver, err := GetVersion(helperVersionCommander{})
+	// OpenVASScanner instance
+	ovas := NewOpenVASScanner()
+
+	ver, err := ovas.GetVersion(helperVersionCommander{})
 	if err != nil {
 		t.Fatalf("Error: Unable to get Version")
 	}
@@ -221,7 +209,10 @@ func TestGetVersion(t *testing.T) {
 
 // TestGetSettings tests if the information getting from the openvas settings is extracted correctly
 func TestGetSettings(t *testing.T) {
-	set, err := GetSettings(helperSettingsCommander{})
+	// OpenVASScanner instance
+	ovas := NewOpenVASScanner()
+
+	set, err := ovas.GetSettings(helperSettingsCommander{})
 	fmt.Printf("%v\n", set)
 
 	if err != nil {